move ExactNumeric to common
