+++ sigmastate/src/main/scala/sigmastate/eval/CompiletimeCosting.scala
@@ -12,7 +12,7 @@ import sigmastate.SCollection._
 import sigmastate.SBigInt._
 import sigmastate.Values.Value.Typed
 import sigmastate.lang.Terms
-import sigma.util.Extensions._
+import scalan.util.Extensions._
 
 trait CompiletimeCosting extends RuntimeCosting { IR: IRContext =>
   import builder._
