+++ sigmastate/src/main/scala/sigmastate/serialization/trees/Relation2Serializer.scala
@@ -6,7 +6,7 @@ import sigmastate.serialization.OpCodes._
 import sigmastate.serialization.ValueSerializer
 import sigmastate.utils.{SigmaByteReader, SigmaByteWriter}
 import sigmastate.serialization.ValueSerializer._
-import sigma.util.Extensions._
+import scalan.util.Extensions._
 
 case class Relation2Serializer[S1 <: SType, S2 <: SType, R <: Value[SBoolean.type]]
 (override val opDesc: RelationCompanion,
