+++ sigmastate/src/main/scala/sigmastate/types.scala
@@ -10,7 +10,7 @@ import scalan.RType.GeneralType
 import sigmastate.SType.{TypeCode, AnyOps}
 import sigmastate.interpreter.CryptoConstants
 import sigmastate.utils.Overloading.Overload1
-import sigma.util.Extensions._
+import scalan.util.Extensions._
 import sigmastate.SBigInt.MaxSizeInBytes
 import sigmastate.Values._
 import sigmastate.lang.Terms._
